Restaurant And Food lInking

+++ src/app/api/logout/route.js
@@ -1,11 +1,15 @@
 const { cookies } = require('next/headers');
 import { NextResponse } from 'next/server';
 import { cookiesParse } from '../../../../utils/cookies';
+import BadRequestError from '../../../../errors/badRequestErrror';
 const Token = require('../../../../models/token');
 
 export async function GET(request) {
-  const isAuthorized = await cookiesParse(request);
-  console.log('AUTh', isAuthorized._id); 
+  const isAuthorized = await cookiesParse(request)
+  
+  if(!isAuthorized){
+    return BadRequestError('Not Logged In')
+  }
   const tokenDelete = await Token.findOneAndDelete({ user: isAuthorized._id });
   cookies().delete('refreshToken');
   cookies().delete('accessToken');
